refactor(FE-11): useStopForm hook is divided into several separate smaller hooks.
