+++ src/features/stops/model/index.ts
@@ -1,7 +1,7 @@
 export { calculateTripDays } from './calculateTripDays';
 export { createStop } from './createStop';
 
-export { useStopFormViewModel } from './stop-form/useStopFormViewModel';
+export { useStopForm } from './stop-form/useStopForm';
 export { useStopItem } from './stop-item/useStopItem';
 export { useDragAndDrop } from './stop-item/useDragAndDrop';
 export { useStopMarker } from './stop-marker/useStopMarker';
