+++ src/features/stops/model/stop-item/useStopItem.ts
@@ -1,17 +1,32 @@
 import { useCallback } from 'react';
 import { useNavigate } from 'react-router';
 
+import { Stop } from 'entities/stop';
+import { Trip } from 'entities/trip';
 import { useMap } from 'features/map';
 import { removeStop } from 'features/trip';
 import { useAppDispatch } from 'shared/lib';
 
-import { UseStopItemHookResult, UseStopItemProps } from './types';
+export interface UseStopItemProps {
+	tripId: Trip['id'];
+	stop: Stop;
+	day?: string;
+}
+
+export interface UseStopItemResult {
+	stopData: Stop;
+	displayDay?: string;
+
+	onDeleteClick: () => void;
+	onEditClick: () => void;
+	onItemClick: () => void;
+}
 
 export const useStopItem = ({
 	stop,
 	tripId,
 	day,
-}: UseStopItemProps): UseStopItemHookResult => {
+}: UseStopItemProps): UseStopItemResult => {
 	const dispatch = useAppDispatch();
 	const navigate = useNavigate();
 	const { map } = useMap();
