--- src/features/stops/model/stop-marker/types.ts
@@ -1,10 +1,0 @@
-import { Holiday } from 'entities/holidays';
-import { Stop } from 'entities/stop';
-import { Weather } from 'entities/weather';
-
-export interface UseStopMarkerHookResult {
-	weather: Weather | null;
-	holiday: Holiday | null;
-	isLoading: boolean;
-	updateStopInfo: (stop: Stop) => Promise<void>;
-}
